change the display of notifications

+++ resources/views/layout.blade.php
@@ -34,50 +34,54 @@
 			<h2 class="notification__title">NOTIFICATIONS</h2>
 			<ul class="notification__group--box">
 				@if ( count($notifications) != 'null' )
+					<?php $nNotification = 0; ?>
 					@foreach ($notifications as $not)
-						<li class="notification__item">
-							@if ($not->not_context == 1)
+						@if( $nNotification < 3 )
+							<li class="notification__item">
+								@if ($not->not_context == 1)
 
-							@endif
-
-							@if ($not->not_context == 2)
-
-							@endif
-
-							@if ($not->not_context == 3)
-
-							@endif
-
-							@if ($not->not_context == 4)
-
-							@endif
-
-							@if ($not->not_context == 5)
-								<div class="notification__content notification__color--green">
-									<a href="{{ action('PageController@viewUser', [ 'id' => $not->user_id]) }}"> {{$not->user_firstname}} {{$not->user_name}}</a> Demande accès au cours de <a href="{{ action('CourseController@view', [ 'id' => $not->course_id, 'action' => 1 ]) }}">{{$not->course_title}}</a>
-								</div>
-								@if ($not->not_seen != 3)
-									<div class="notification__actionGroup double">
-										<a class="notification__button icon unlink success" href="{!! action( 'CourseController@acceptStudent', ['id_course' => $not->course_id, 'id_user' => $not->user_id] ) !!}"><span class="icon-check"></span><span class="hidden">Ajouter</span></a>
-
-										<a class="notification__button icon unlink danger" href="{!! action( 'CourseController@removeStudentFromCourse', ['id_course' => $not->course_id, 'id_user' => $not->user_id] ) !!}"><span class="icon-close"></span><span class="hidden">Refuser</span></a>
-									</div>
 								@endif
-							@endif
 
-							@if ($not->not_context == 6)
-								{{$not->not_title}} <a href="{{ action('CourseController@view', [ 'id' => $not->course_id, 'action' => 1 ]) }}">{{$not->course_title}}</a>
-							@endif
+								@if ($not->not_context == 2)
 
-							@if ($not->not_context == 7)
-								{{$not->not_title}} {{$not->course_title}}
-							@endif
+								@endif
 
-							@if ($not->not_context == 8)
-								<a href="{{ action('PageController@viewUser', [ 'id' => $not->user_id]) }}">{{$not->user_name}}</a> {{$not->not_title}} <a href="{{ action('CourseController@view', [ 'id' => $not->course_id, 'action' => 1 ]) }}">{{$not->course_title}}</a>
-							@endif
-							<div class="clear"></div>
-						</li>
+								@if ($not->not_context == 3)
+
+								@endif
+
+								@if ($not->not_context == 4)
+
+								@endif
+
+								@if ($not->not_context == 5)
+									<div class="notification__content notification__color--green">
+										<a href="{{ action('PageController@viewUser', [ 'id' => $not->user_id]) }}"> {{$not->user_firstname}} {{$not->user_name}}</a> Demande accès au cours de <a href="{{ action('CourseController@view', [ 'id' => $not->course_id, 'action' => 1 ]) }}">{{$not->course_title}}</a>
+									</div>
+									@if ($not->not_seen != 3)
+										<div class="notification__actionGroup double">
+											<a class="notification__button icon unlink success" href="{!! action( 'CourseController@acceptStudent', ['id_course' => $not->course_id, 'id_user' => $not->user_id] ) !!}"><span class="icon-check"></span><span class="hidden">Ajouter</span></a>
+
+											<a class="notification__button icon unlink danger" href="{!! action( 'CourseController@removeStudentFromCourse', ['id_course' => $not->course_id, 'id_user' => $not->user_id] ) !!}"><span class="icon-close"></span><span class="hidden">Refuser</span></a>
+										</div>
+									@endif
+								@endif
+
+								@if ($not->not_context == 6)
+									{{$not->not_title}} <a href="{{ action('CourseController@view', [ 'id' => $not->course_id, 'action' => 1 ]) }}">{{$not->course_title}}</a>
+								@endif
+
+								@if ($not->not_context == 7)
+									{{$not->not_title}} {{$not->course_title}}
+								@endif
+
+								@if ($not->not_context == 8)
+									<a href="{{ action('PageController@viewUser', [ 'id' => $not->user_id]) }}">{{$not->user_name}}</a> {{$not->not_title}} <a href="{{ action('CourseController@view', [ 'id' => $not->course_id, 'action' => 1 ]) }}">{{$not->course_title}}</a>
+								@endif
+								<div class="clear"></div>
+							</li>
+							<?php $nNotification++; ?>
+						@endif
 					@endforeach
 				@else
 					<li class="notification__item--null">
